Fix: 채팅창을 닫으면 채팅 연결이 끊기도록 수정
